Refactor delete message

+++ app/src/main/java/com/expirate/expirat/ui/managegroups/ManageGroupsActivity.java
@@ -8,6 +8,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.Snackbar;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
@@ -23,6 +24,7 @@ import com.expirate.expirat.ui.BaseActiviy;
 import com.expirate.expirat.ui.widget.DialogNewGroup;
 
 import java.util.List;
+import java.util.Locale;
 
 import javax.inject.Inject;
 
@@ -97,16 +99,25 @@ public class ManageGroupsActivity extends BaseActiviy implements ManageGroupsCon
 
     @Override
     public void onDeleteClick(long id, String typeName) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setMessage(getString(R.string.delete_message));
+        builder.setPositiveButton(R.string.btn_delete, (dialog, which) -> {
+            tracker.send(new HitBuilders.EventBuilder()
+                    .setCategory("Manage Group")
+                    .setAction("Delete Group")
+                    .setLabel(typeName)
+                    .build());
 
-        tracker.send(new HitBuilders.EventBuilder()
-                .setCategory("Manage Group")
-                .setAction("Delete Group")
-                .setLabel(typeName)
-                .build());
+            presenter.deleteGroup(id);
+            //noinspection ConstantConditions
+            Snackbar.make(getCurrentFocus(),
+                    String.format(Locale.getDefault(), "Delete %s", typeName),
+                    Snackbar.LENGTH_SHORT).show();
 
-        presenter.deleteGroup(id);
-        //noinspection ConstantConditions
-        Snackbar.make(getCurrentFocus(), "Delete 1 item", Snackbar.LENGTH_SHORT).show();
+            dialog.dismiss();
+        });
+        builder.setNegativeButton(R.string.btn_cancel, (dialog, which) -> dialog.dismiss());
+        builder.show();
     }
 
     @Override
@@ -118,6 +129,9 @@ public class ManageGroupsActivity extends BaseActiviy implements ManageGroupsCon
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
+            case android.R.id.home:
+                onBackPressed();
+                return true;
             case R.id.action_add:
                 dialogNewItemGroup();
                 return true;
